+++ src/main/java/com/athenhub/stockservice/stock/infrastructure/client/member/MemberClient.java
@@ -1,7 +1,6 @@
-package com.athenhub.stockservice.stock.infrastructure.client;
+package com.athenhub.stockservice.stock.infrastructure.client.member;
 
 import com.athenhub.stockservice.global.infrastructure.feignclient.FeignClientConfig;
-import com.athenhub.stockservice.stock.infrastructure.dto.MemberInfo;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
