+++ src/main/java/com/athenhub/stockservice/stock/infrastructure/client/member/MemberRole.java
@@ -1,4 +1,4 @@
-package com.athenhub.stockservice.stock.infrastructure;
+package com.athenhub.stockservice.stock.infrastructure.client.member;
 
 public enum MemberRole {
   MASTER_MANAGER,
