+++ src/main/java/com/athenhub/stockservice/stock/infrastructure/client/product/DefaultProductAccessPermissionValidator.java
@@ -1,9 +1,7 @@
-package com.athenhub.stockservice.stock.infrastructure;
+package com.athenhub.stockservice.stock.infrastructure.client.product;
 
 import com.athenhub.stockservice.stock.domain.dto.AccessContext;
 import com.athenhub.stockservice.stock.domain.service.ProductAccessPermissionValidator;
-import com.athenhub.stockservice.stock.infrastructure.client.ProductClient;
-import com.athenhub.stockservice.stock.infrastructure.client.ProductDetail;
 import java.util.Objects;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
